Fix install script with dot-path

+++ install.sh
@@ -3,6 +3,12 @@
 DIR=`dirname $0`
 FILES="bash_aliases bash_logout bash_profile bashrc vimrc"
 
+if [ $DIR = "." ]
+then
+	echo "Please invoke with full path."
+	exit 1
+fi
+
 for f in $FILES
 do
 	rm -iv ~/.$f
